actions/test_ncWMS_speed/requirements.txt
aiohttp
xmltodict
pyyaml
sqlachemy
modelmeta
